update a big change

+++ Models/PagePagination.cs
@@ -8,7 +8,7 @@ namespace FinalWebApp.Models
 {
     public class PagePagination<TData>
     {
-        public int TotalElements { get; set; }
+        public long TotalElements { get; set; }
         public int First { get; set; }
         public int Last { get; set; }
         public IEnumerable<TData> Content { get; set; }
